Update RSpecs for active/discontinued, update slicing serialization to only active_services
